requirements.txt
pyserial
pymavlink
ipython
black
tdqm
prompt_toolkit
colorama
